% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R library summary}
\usage{
lib_summary(sizes = FALSE)
}
\arguments{
\item{sizes}{A logical indicating whether or not to add the space size (in bytes)}
}
\value{
A data.frame containing the count of packages in each of your libraries.
}
\description{
R library summary
}
\examples{
\dontrun{
lib_summary()
}
lib_summary(sizes = TRUE)
}
