% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer2-package.R
\docType{package}
\name{libminer2-package}
\alias{libminer2}
\alias{libminer2-package}
\title{libminer2: Explore R Libraries}
\description{
Provides functions for learning about your R libraries, and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MonkmanMH/libminer2}
  \item Report bugs at \url{https://github.com/MonkmanMH/libminer2/issues}
}

}
\author{
\strong{Maintainer}: Martin Monkman \email{monkmanmh@yahoo.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
