% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminer2-package.R
\docType{package}
\name{libminer2-package}
\alias{libminer2}
\alias{libminer2-package}
\title{libminer2: Explore Your R Libraries}
\description{
Provides functions for learning about your R libraries, and the pacakges you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ateucher/libminer2}
  \item Report bugs at \url{https://github.com/ateucher/libminer2/issues}
}

}
\author{
\strong{Maintainer}: Andy Teucher \email{andy.teucher@gmail.com} (\href{https://orcid.org/0000-0002-7840-692X}{ORCID})

}
\keyword{internal}
